Add onchain types for version oracle script (#1012)

+++ toolkit/smart-contracts/offchain/src/reserve/mod.rs
@@ -41,7 +41,7 @@ impl ReserveData {
 	) -> Result<Self, anyhow::Error> {
 		let version_oracle = scripts_data::version_oracle(genesis_utxo, ctx.network)?;
 		let auth_policy_version_utxo = find_script_utxo(
-			raw_scripts::ScriptId::ReserveAuthPolicy as u32,
+			raw_scripts::ScriptId::ReserveAuthPolicy,
 			&version_oracle,
 			ctx,
 			client,
@@ -52,18 +52,14 @@ impl ReserveData {
 				"Reserve Auth Version Utxo not found, is the Reserve Token Management initialized?"
 			)
 		})?;
-		let validator_version_utxo = find_script_utxo(
-			raw_scripts::ScriptId::ReserveValidator as u32,
-			&version_oracle,
-			ctx,
-			client,
-		)
-		.await?
-		.ok_or_else(|| {
-			anyhow!(
-				"Reserve Validator Version Utxo not found, is the Reserve Token Management initialized?"
-			)
-		})?;
+		let validator_version_utxo =
+			find_script_utxo(raw_scripts::ScriptId::ReserveValidator, &version_oracle, ctx, client)
+				.await?
+				.ok_or_else(|| {
+					anyhow!(
+						"Reserve Validator Version Utxo not found, is the Reserve Token Management initialized?"
+					)
+				})?;
 		let scripts = scripts_data::reserve_scripts(genesis_utxo, ctx.network)?;
 		Ok(ReserveData { scripts, auth_policy_version_utxo, validator_version_utxo })
 	}
